Refactor implementation of DataConnector for s3/ftp into ListingTableConnector
