+++ crates/runtime/src/dataconnector/ftp.rs
@@ -14,22 +14,15 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-use crate::object_store_registry::macros::impl_listing_data_connector;
-use datafusion::datasource::file_format::csv::CsvFormat;
-use datafusion::datasource::file_format::file_compression_type::FileCompressionType;
-use datafusion::datasource::file_format::parquet::ParquetFormat;
-use datafusion::datasource::file_format::FileFormat;
-use datafusion::error::DataFusionError;
 use secrets::{AnyErrorResult, Secret};
 use snafu::prelude::*;
 use spicepod::component::dataset::Dataset;
 use std::pin::Pin;
-use std::str::FromStr;
 use std::sync::Arc;
 use std::{collections::HashMap, future::Future};
 use url::{form_urlencoded, Url};
 
-use super::{DataConnector, DataConnectorFactory};
+use super::{DataConnector, DataConnectorFactory, ListingTableConnector};
 
 #[derive(Debug, Snafu)]
 pub enum Error {
@@ -37,39 +30,6 @@ pub enum Error {
     UnableToParseURL {
         url: String,
         source: url::ParseError,
-    },
-
-    #[snafu(display("{source}"))]
-    UnableToGetReadProvider {
-        source: Box<dyn std::error::Error + Send + Sync>,
-    },
-
-    #[snafu(display("{source}"))]
-    UnableToGetReadWriteProvider {
-        source: Box<dyn std::error::Error + Send + Sync>,
-    },
-
-    #[snafu(display("{source}"))]
-    UnableToBuildObjectStore {
-        source: object_store::Error,
-    },
-
-    ObjectStoreNotImplemented,
-
-    #[snafu(display("{source}"))]
-    UnableToBuildLogicalPlan {
-        source: DataFusionError,
-    },
-
-    #[snafu(display("Unsupported file format {format} in S3 Connector"))]
-    UnsupportedFileFormat {
-        format: String,
-    },
-
-    #[snafu(display("Unsupported compression type for CSV"))]
-    UnsupportedCompressionType {
-        source: DataFusionError,
-        compression_type: String,
     },
 }
 
@@ -93,7 +53,11 @@ impl DataConnectorFactory for FTP {
     }
 }
 
-impl FTP {
+impl ListingTableConnector for FTP {
+    fn get_params(&self) -> &HashMap<String, String> {
+        &self.params
+    }
+
     fn get_object_store_url(&self, dataset: &Dataset) -> AnyErrorResult<Url> {
         let mut fragments = vec![];
         let mut fragment_builder = form_urlencoded::Serializer::new(String::new());
@@ -122,44 +86,6 @@ impl FTP {
 
         Ok(ftp_url)
     }
-
-    fn get_file_format_and_extension(&self) -> AnyErrorResult<(Arc<dyn FileFormat>, String)> {
-        let params = &self.params;
-        let extension = params.get("file_extension").cloned();
-
-        match params.get("file_format").map(String::as_str) {
-            Some("csv") => Ok((
-                get_csv_format(params)?,
-                extension.unwrap_or(".csv".to_string()),
-            )),
-            None | Some("parquet") => Ok((
-                Arc::new(ParquetFormat::default()),
-                extension.unwrap_or(".parquet".to_string()),
-            )),
-            Some(format) => Err(Error::UnsupportedFileFormat {
-                format: format.to_string(),
-            }
-            .into()),
-        }
-    }
-}
-
-fn get_csv_format(params: &HashMap<String, String>) -> AnyErrorResult<Arc<CsvFormat>> {
-    let compression_type = params.get("compression_type").map_or("", |f| f);
-    let has_header = params.get("has_header").map_or(true, |f| f == "true");
-    let delimiter = params
-        .get("delimiter")
-        .map_or(b',', |f| *f.as_bytes().first().unwrap_or(&b','));
-
-    Ok(Arc::new(
-        CsvFormat::default()
-            .with_has_header(has_header)
-            .with_file_compression_type(
-                FileCompressionType::from_str(compression_type)
-                    .context(UnsupportedCompressionTypeSnafu { compression_type })?,
-            )
-            .with_delimiter(delimiter),
-    ))
 }
 
 pub(crate) fn get_secret_or_param(
@@ -185,5 +111,3 @@ pub(crate) fn get_secret_or_param(
 
     None
 }
-
-impl_listing_data_connector!(FTP);
