+++ crates/runtime/src/object_store_registry.rs
@@ -115,85 +115,3 @@ pub(crate) fn default_runtime_env() -> Arc<RuntimeEnv> {
         .unwrap_or_default(),
     )
 }
-
-pub mod macros {
-    macro_rules! impl_listing_data_connector {
-        ($connector:ty) => {
-            use crate::object_store_registry::default_runtime_env;
-            use async_trait::async_trait;
-            use datafusion::datasource::listing::{
-                ListingOptions, ListingTable, ListingTableConfig, ListingTableUrl,
-            };
-            use datafusion::datasource::TableProvider;
-            use datafusion::execution::config::SessionConfig;
-            use datafusion::execution::context::SessionContext;
-            use std::any::Any;
-
-            #[async_trait]
-            impl DataConnector for $connector {
-                fn as_any(&self) -> &dyn Any {
-                    self
-                }
-
-                async fn read_provider(
-                    &self,
-                    dataset: &Dataset,
-                ) -> super::DataConnectorResult<Arc<dyn TableProvider>> {
-                    let ctx = SessionContext::new_with_config_rt(
-                        SessionConfig::new().set_bool(
-                            "datafusion.execution.listing_table_ignore_subdirectory",
-                            false,
-                        ),
-                        default_runtime_env(),
-                    );
-
-                    let url = self.get_object_store_url(dataset).context(
-                        super::InvalidConfigurationSnafu {
-                            dataconnector: stringify!($connector),
-                            message: "Unable to parse URL",
-                        },
-                    )?;
-
-                    let table_path = ListingTableUrl::parse(url).boxed().context(
-                        super::InvalidConfigurationSnafu {
-                            dataconnector: stringify!($connector).to_string(),
-                            message: "Unable to parse URL",
-                        },
-                    )?;
-
-                    let (file_format, extension) = self.get_file_format_and_extension().context(
-                        super::InvalidConfigurationSnafu {
-                            dataconnector: stringify!($connector).to_string(),
-                            message: "Unable to resolve file_format and file_extension",
-                        },
-                    )?;
-                    let options = ListingOptions::new(file_format).with_file_extension(&extension);
-
-                    let resolved_schema = options
-                        .infer_schema(&ctx.state(), &table_path)
-                        .await
-                        .boxed()
-                        .context(super::InvalidConfigurationSnafu {
-                            dataconnector: stringify!($connector).to_string(),
-                            message: "Unable to infer files schema",
-                        })?;
-
-                    let config = ListingTableConfig::new(table_path)
-                        .with_listing_options(options)
-                        .with_schema(resolved_schema);
-
-                    let table = ListingTable::try_new(config).boxed().context(
-                        super::InvalidConfigurationSnafu {
-                            dataconnector: stringify!($connector).to_string(),
-                            message: "Unable to list files in data provider",
-                        },
-                    )?;
-
-                    Ok(Arc::new(table))
-                }
-            }
-        };
-    }
-
-    pub(crate) use impl_listing_data_connector;
-}
